Create RemoveDuplicateLocalAdmin.ps1 (#25)
* Create RemoveDuplicateLocalAdmin.ps1

This script identifies duplicate reg keys in the profilelist and then queries the local SAM database to identify the SID of the true local admin profile. It then removes any duplicates. The duplicate regkeys have consistently caused failures when performing windows 11 upgrades and this should resolve this issue.

* Updated script to cleanup all duplicate local users

* Delete msft-windows/RemoveDuplicateLocalAdmin.ps1

Replacing this with better script

* Add files via upload

* Update RemoveDuplicateLocalAdmin.ps1

Added in script template